Remove dir
Added a check if the $file in the delete function is a dir or not, if it
is a dir then use rmdir() otherwise use unlink()

+++ ftp.php
@@ -141,10 +141,17 @@
 		}
 
 
-		# Unlink the file (unlink will make the system forget about the file, aka delete it)!
+		if(is_dir($folderPath . "/" . $file)
+		{
+			rmdir($folderPath . "/" . $file);
+		}
+		else
+		{
+ 			# Unlink the file (unlink will make the system forget about the file, aka delete it)!
 
-		unlink($folderPath . "/" . $file);
-
+			unlink($folderPath . "/" . $file);
+		}
+		
 		echo '<div class="alert alert-success"> SUCCESS!!! <br>';
 		echo 'Deleted ' . $file;
 		echo '</div>';
